Fixed navigation and disabled reference and udstyr and hiding tilsyn
Previously the navigation in stamdata would not work correctly if a user navigated between the tabs and then selecting to change the tidsserie. It would remove the ?page and only display the &tab and if the page was refreshed it would go back to pejling page, but the url still had &tab. Reference is now disabled if the tstype is anything except for vandstand. Udstyr tab is disabled for calculated tidsserie.
Tilsyn is hidden for calculated tidsserie