Added event for usesearchparams
Since Navigation is unsupported in some browser and including Safari, we cant use it. Therefore I changed to adding an eventlistener to the Document for when a click event happens. In this case it will fire every time a click happens on the page. This is not the ideal solution, and should be changed if a better solution is found