Further updates to editstamdata and opretstamdata
Date validation for when selecting a start date that is later than end date = error and selecting an enddate that is earlier than start date = error.
Opret stamdata from navbar when on a location will always navigate to timeseries tab
Edit stamdata now allows the user to add a unit if none were added under the creation of a timeseries.
Initial project no is now disabled when selecting a location in opretstamdata