Implemented parking
New parking functionality implemented. Right clicking on map will open a context menu which has an option to create a new parking.
Right clicking on a location will open a menu where it is possible to either create a parking attached to location or attaching a parking to location
Highligting a location will as well highlight the parking attached to the location.