Implemented transferlist
The vision of this feature is to be able to assign which Ressources should be added to the car before a service is done. On desktop its displayed as two lists with buttons in between. Checking off the checkboxes in one list will enable either the move to left or move to right button depending on which list was interacted with. On mobile its a multiselect component just like how location groups is.