Fixes WAT-484
Errormessages will show when trying to use "gem & afslut" as a toast as well as an icon on the tab.
When using "videre til X" it will trigger validation on the tab it is pressed from and show an icon on the tab as well as highlighting the error on the fields.
Fixed error with timeseries tab not being dirty when changing timeseries type.